Extract find composite slo definitions
