+++ x-pack/solutions/observability/plugins/slo/server/routes/composite_slo/find_composite_slo.ts
@@ -5,105 +5,9 @@
  * 2.0.
  */
 
-import type { QueryDslQueryContainer } from '@elastic/elasticsearch/lib/api/types';
-import type { ElasticsearchClient } from '@kbn/core/server';
-import type { Paginated } from '@kbn/slo-schema';
 import { findCompositeSLOParamsSchema } from '@kbn/slo-schema';
-import { COMPOSITE_SUMMARY_INDEX_NAME } from '../../../common/constants';
-import type { CompositeSLODefinition } from '../../domain/models';
-import type { CompositeSLORepository } from '../../services';
+import { findCompositeSloDefinitions } from '../../services/composites/find_composite_slo_definitions';
 import { createCompositeSloServerRoute } from './create_composite_slo_server_route';
-
-interface FindCompositeSloParams {
-  spaceId: string;
-  search: string | undefined;
-  tags: string[];
-  statusFilter: string[];
-  sortBy: 'name' | 'createdAt' | 'updatedAt' | undefined;
-  sortDirection: 'asc' | 'desc' | undefined;
-  page: number;
-  perPage: number;
-}
-
-interface Dependencies {
-  compositeRepository: CompositeSLORepository;
-  esClient: ElasticsearchClient;
-}
-
-/**
- * Reads from the composite summary index for filter/sort/search/paginate, then looks up the
- * full SO definitions for the page's IDs. The summary doc denormalises the fields needed for
- * filtering (spaceId, name, tags, status) but the response shape is the SO definition, so we
- * always lookup definitions for the page (bounded by perPage).
- *
- * Composites without a persisted summary doc (e.g. a still-bootstrapping deployment from a
- * pre-inline-persist build) will not appear here until the background task runs.
- */
-async function findCompositeSlos(
-  {
-    spaceId,
-    search,
-    tags,
-    statusFilter,
-    sortBy,
-    sortDirection,
-    page,
-    perPage,
-  }: FindCompositeSloParams,
-  { compositeRepository, esClient }: Dependencies
-): Promise<Paginated<CompositeSLODefinition>> {
-  const filters: QueryDslQueryContainer[] = [{ term: { spaceId } }];
-  if (search) {
-    filters.push({
-      simple_query_string: {
-        query: search,
-        fields: ['compositeSlo.name'],
-        default_operator: 'AND',
-      },
-    });
-  }
-  if (tags.length > 0) {
-    filters.push({ terms: { 'compositeSlo.tags': tags } });
-  }
-  if (statusFilter.length > 0) {
-    filters.push({ terms: { status: statusFilter } });
-  }
-
-  const sortField =
-    sortBy === 'name' ? 'compositeSlo.name.keyword' : `compositeSlo.${sortBy ?? 'createdAt'}`;
-  const order = sortDirection ?? 'desc';
-
-  const response = await esClient.search<{ compositeSlo: { id: string } }>({
-    index: COMPOSITE_SUMMARY_INDEX_NAME,
-    from: (page - 1) * perPage,
-    size: perPage,
-    track_total_hits: true,
-    query: { bool: { filter: filters } },
-    sort: [{ [sortField]: { order } }],
-    _source: ['compositeSlo.id'],
-  });
-
-  const total =
-    typeof response.hits.total === 'number' ? response.hits.total : response.hits.total?.value ?? 0;
-
-  const compositeIds = response.hits.hits
-    .map((hit) => hit._source?.compositeSlo.id)
-    .filter((id): id is string => typeof id === 'string');
-
-  if (compositeIds.length === 0) {
-    return { page, perPage, total, results: [] };
-  }
-
-  const definitions = await compositeRepository.findAllByIds(compositeIds);
-  const definitionsById = new Map(definitions.map((d) => [d.id, d]));
-
-  // Preserve the order returned by the ES search (which honours the sort).
-  const results = compositeIds
-    .map((id) => definitionsById.get(id))
-    .filter((d): d is CompositeSLODefinition => d !== undefined);
-
-  return { page, perPage, total, results };
-}
 
 export const findCompositeSLORoute = createCompositeSloServerRoute({
   endpoint: 'GET /api/observability/slo_composites 2023-10-31',
@@ -114,7 +18,7 @@ export const findCompositeSLORoute = createCompositeSloServerRoute({
     },
   },
   params: findCompositeSLOParamsSchema,
-  handler: async ({ context, params, logger, request, plugins, getScopedClients }) => {
+  handler: async ({ params, logger, request, getScopedClients }) => {
     const { scopedClusterClient, compositeRepository, spaceId } = await getScopedClients({
       request,
       logger,
@@ -126,7 +30,7 @@ export const findCompositeSLORoute = createCompositeSloServerRoute({
     const tags = query.tags ? query.tags.split(',').map((tag) => tag.trim()) : [];
     const statusFilter = query.status ?? [];
 
-    return await findCompositeSlos(
+    return await findCompositeSloDefinitions(
       {
         spaceId,
         search: query.search,
